refactor(fileUtils): promote resolveGalleryImage helper, dedupe 5 sites
Extract the basename + resolvePath + startsWith(imagesRoot) gallery-path
guard into a single `resolveGalleryImage(name, { mustExist })` helper in
server/lib/fileUtils.js. Replace inline implementations in videoGen,
imageGen, visualStages, sceneRunner, and imageGen/local. The four
non-videoGen sites now pick up the `isFile()` check too, so the
images-root directory itself can't slip through as an "image path".

Removed the now-done PLAN.md item under "Code quality / dedup".
